+++ src/main/java/com/cburch/logisim/file/XmlWriter.java
@@ -441,7 +441,7 @@ class XmlWriter {
       Integer mods = entry.getKey();
       Tool tool = entry.getValue();
       Element toolElt = fromTool(tool);
-      String mapValue = InputEventUtil.toString(mods.intValue());
+      String mapValue = InputEventUtil.toString(mods);
       toolElt.setAttribute("map", mapValue);
       elt.appendChild(toolElt);
     }
