+++ src/main/java/com/cburch/logisim/std/gates/XorGate.java
@@ -105,7 +105,7 @@ class XorGate extends AbstractGate {
     Object behavior = attrs.getValue(GateAttributes.ATTR_XOR);
     if (behavior == GateAttributes.XOR_ODD) {
       Object inputs = attrs.getValue(GateAttributes.ATTR_INPUTS);
-      if (inputs == null || ((Integer) inputs).intValue() != 2) {
+      if (inputs == null || (Integer) inputs != 2) {
         isOdd = true;
       }
     }
