+++ src/main/java/com/cburch/logisim/std/memory/CounterAttributes.java
@@ -116,7 +116,7 @@ class CounterAttributes extends AbstractAttributeSet {
       int oldW = oldWidth.getWidth();
       int newW = newWidth.getWidth();
       Long oldValObj = base.getValue(Counter.ATTR_MAX);
-      long oldVal = oldValObj.longValue();
+      long oldVal = oldValObj;
       base.setValue(StdAttr.WIDTH, newWidth);
       if (newW > oldW) {
         newMax = newWidth.getMask();
@@ -130,9 +130,9 @@ class CounterAttributes extends AbstractAttributeSet {
       }
       fireAttributeValueChanged(StdAttr.WIDTH, newWidth, null);
     } else if (attr == Counter.ATTR_MAX) {
-      long oldVal = base.getValue(Counter.ATTR_MAX).longValue();
+      long oldVal = base.getValue(Counter.ATTR_MAX);
       BitWidth width = base.getValue(StdAttr.WIDTH);
-      long newVal = ((Long) value).longValue() & width.getMask();
+      long newVal = (Long) value & width.getMask();
       if (newVal != oldVal) {
         value = (V) Long.valueOf(newVal);
       }
