+++ src/main/java/com/cburch/logisim/std/memory/Ram.java
@@ -99,7 +99,7 @@ public class Ram extends Mem {
     public Value getLogValue(InstanceState state, Object option) {
       if (option instanceof Long) {
         MemState s = (MemState) state.getData();
-        long addr = ((Long) option).longValue();
+        long addr = (Long) option;
         return Value.createKnown(BitWidth.create(s.getDataBits()), s.getContents().get(addr));
       } else {
         return Value.NIL;
