+++ src/main/java/com/cburch/logisim/std/wiring/ClockHDLGeneratorFactory.java
@@ -236,9 +236,9 @@ public class ClockHDLGeneratorFactory extends AbstractHDLGeneratorFactory {
   public SortedMap<String, Integer> GetParameterMap(
       Netlist Nets, NetlistComponent ComponentInfo, FPGAReport Reporter) {
     SortedMap<String, Integer> ParameterMap = new TreeMap<>();
-    int HighTicks = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_HIGH).intValue();
-    int LowTicks = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_LOW).intValue();
-    int Phase = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_PHASE).intValue();
+    int HighTicks = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_HIGH);
+    int LowTicks = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_LOW);
+    int Phase = ComponentInfo.GetComponent().getAttributeSet().getValue(Clock.ATTR_PHASE);
     Phase = Phase % (HighTicks + LowTicks);
     int MaxValue = Math.max(HighTicks, LowTicks);
     int nr_of_bits = 0;
